wip: architecture for primop lifting

+++ src/types.rs
@@ -88,7 +88,7 @@ impl Type {
 
 /// arrow types have arity of whatever their return type is, plus 1.
 /// all others have arity 0 since they are not functions.
-pub fn type_arity(ty: Type) -> u8 {
+pub fn type_arity(ty: Type) -> usize {
     match ty {
         Type::TArr(_arg, ret) => 1 + type_arity(*ret),
         _ => 0,
